Upgrade functions to ft printf efficienty level

+++ srcs/ft_abs.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_abs.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmougino <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nmougino <nmougino@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/03/15 17:59:01 by nmougino          #+#    #+#             */
-/*   Updated: 2016/03/15 18:13:12 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/08/04 17:27:48 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_abs(int nb)
+#include "libft.h"
+
+uintmax_t	ft_abs(intmax_t nb)
 {
 	if (nb < 0)
 		return (-nb);
